refactored code to be modular
