+++ CMakeLists.txt
@@ -46,7 +46,13 @@ else ()
 endif ()
 
 function(setup_ygm_target exe_name)
-	add_executable(${exe_name} ${exe_name}.cpp)
+	# First argument is the executable name, rest are additional source files
+	set(SOURCES ${exe_name}.cpp)
+	foreach(ARG ${ARGN})
+		list(APPEND SOURCES ${ARG})
+	endforeach()
+	
+	add_executable(${exe_name} ${SOURCES})
 	target_link_libraries(${exe_name} PRIVATE ygm::ygm)
 	target_include_directories(${exe_name} PRIVATE "${PROJECT_SOURCE_DIR}/include")
 	target_include_directories(${exe_name} PRIVATE ${PROJECT_SOURCE_DIR}/include ${Boost_INCLUDE_DIRS})
