+++ src/CMakeLists.txt
@@ -1,3 +1,3 @@
 setup_ygm_target(scc-sequential)
 setup_ygm_target(scc-parallel)
-setup_ygm_target(scc-ygm)
+setup_ygm_target(scc-ygm graph.cpp scc.cpp)
